google-cloud-functions/extract-text-from-docx/requirements.txt
functions-framework==3.*
python-docx
requests
